requirements.txt
# =========================================
# Google API (Sheets, Drive, Gmail)
# =========================================
gspread>=6.0.2
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.129.0

# =========================================
# RSS Parsing
# =========================================
feedparser>=6.0.10

# =========================================
# HTTP Requests
# =========================================
requests>=2.31.0

# =========================================
# Scheduling
# =========================================
APScheduler>=3.10.4

# =========================================
# Image Processing
# =========================================
Pillow>=11.0.0

# =========================================
# Testing
# =========================================
pytest>=8.2.0
pytest-cov>=5.0.0

# =========================================
# Utilities
# =========================================
python-dotenv>=1.0.1
pytz>=2024.1
